added new algorithm Finder to route-finder, remove old Dijkstra algorithm
