+++ pkg/route-finder/store/mark_and_sweep.go
@@ -44,7 +44,6 @@ type Graph struct {
 	visited map[cipher.PubKey]*vertex
 	graph   map[cipher.PubKey]*vertex
 	dist    map[*vertex]int
-	hops    map[*vertex]int
 	prev    map[*vertex]*vertex
 }
